observables: refactoring NetworkServer
