+++ scalastyle-config.xml
@@ -81,7 +81,7 @@
     </check>
     <check level="warning" class="org.scalastyle.scalariform.MethodLengthChecker" enabled="true">
         <parameters>
-            <parameter name="maxLength"><![CDATA[100]]></parameter>
+            <parameter name="maxLength"><![CDATA[300]]></parameter>
         </parameters>
     </check>
     <check level="warning" class="org.scalastyle.scalariform.MethodNamesChecker" enabled="true">
