+++ src/main/scala/com/wavesplatform/network/NetworkServer.scala
@@ -4,11 +4,9 @@ import java.net.{InetSocketAddress, NetworkInterface}
 import java.nio.channels.ClosedChannelException
 import java.util.concurrent.ConcurrentHashMap
 
-import com.wavesplatform.features.FeatureProvider
 import com.wavesplatform.metrics.Metrics
-import com.wavesplatform.mining.Miner
+import com.wavesplatform.network.MessageObserver.Messages
 import com.wavesplatform.settings._
-import com.wavesplatform.state2._
 import com.wavesplatform.{UtxPool, Version}
 import io.netty.bootstrap.{Bootstrap, ServerBootstrap}
 import io.netty.channel._
@@ -24,92 +22,117 @@ import monix.reactive.subjects.ConcurrentSubject
 import org.asynchttpclient.netty.channel.NoopHandler
 import org.influxdb.dto.Point
 import scorex.transaction._
-import scorex.utils.{ScorexLogging, Time}
+import scorex.utils.ScorexLogging
 
 import scala.collection.JavaConverters._
 import scala.concurrent.duration._
 
-class NetworkServer(checkpointService: CheckpointService,
-                    blockchainUpdater: BlockchainUpdater,
-                    time: Time,
-                    miner: Miner,
-                    stateReader: StateReader,
-                    settings: WavesSettings,
-                    history: NgHistory,
-                    utxPool: UtxPool,
-                    peerDatabase: PeerDatabase,
-                    allChannels: ChannelGroup,
-                    peerInfo: ConcurrentHashMap[Channel, PeerInfo],
-                    featureProvider: FeatureProvider) extends ScorexLogging {
+trait NS {
+  def connect(remoteAddress: InetSocketAddress): Unit
 
-  @volatile
-  private var shutdownInitiated = false
+  def shutdown(): Unit
 
-  private val bossGroup = new NioEventLoopGroup(0, new DefaultThreadFactory("nio-boss-group", true))
-  private val workerGroup = new NioEventLoopGroup(0, new DefaultThreadFactory("nio-worker-group", true))
-  private val handshake =
-    Handshake(Constants.ApplicationName + settings.blockchainSettings.addressSchemeCharacter, Version.VersionTuple,
+  val messages: Messages
+  val closedChannels: Observable[Channel]
+}
+
+object NetworkServer extends ScorexLogging {
+
+  def apply(settings: WavesSettings,
+            lastBlockInfos: Observable[LastBlockInfo],
+            history: NgHistory,
+            utxPool: UtxPool,
+            peerDatabase: PeerDatabase,
+            allChannels: ChannelGroup,
+            peerInfo: ConcurrentHashMap[Channel, PeerInfo]
+           ): NS = {
+    @volatile var shutdownInitiated = false
+
+    val bossGroup = new NioEventLoopGroup(0, new DefaultThreadFactory("nio-boss-group", true))
+    val workerGroup = new NioEventLoopGroup(0, new DefaultThreadFactory("nio-worker-group", true))
+    val handshake = Handshake(Constants.ApplicationName + settings.blockchainSettings.addressSchemeCharacter, Version.VersionTuple,
       settings.networkSettings.nodeName, settings.networkSettings.nonce, settings.networkSettings.declaredAddress)
 
-  private val trafficWatcher = if (settings.metrics.enable) new TrafficWatcher else new NoopHandler
-  private val messageCodec = new MessageCodec(peerDatabase)
+    val trafficWatcher = if (settings.metrics.enable) new TrafficWatcher else new NoopHandler
+    val messageCodec = new MessageCodec(peerDatabase)
 
-  private val excludedAddresses: Set[InetSocketAddress] = {
-    val localAddresses = if (settings.networkSettings.bindAddress.getAddress.isAnyLocalAddress) {
-      NetworkInterface.getNetworkInterfaces.asScala
-        .flatMap(_.getInetAddresses.asScala
-          .map(a => new InetSocketAddress(a, settings.networkSettings.bindAddress.getPort)))
-        .toSet
-    } else Set(settings.networkSettings.bindAddress)
+    val excludedAddresses: Set[InetSocketAddress] = {
+      val localAddresses = if (settings.networkSettings.bindAddress.getAddress.isAnyLocalAddress) {
+        NetworkInterface.getNetworkInterfaces.asScala
+          .flatMap(_.getInetAddresses.asScala
+            .map(a => new InetSocketAddress(a, settings.networkSettings.bindAddress.getPort)))
+          .toSet
+      } else Set(settings.networkSettings.bindAddress)
 
-    localAddresses ++ settings.networkSettings.declaredAddress.toSet
-  }
+      localAddresses ++ settings.networkSettings.declaredAddress.toSet
+    }
 
-  private val lengthFieldPrepender = new LengthFieldPrepender(4)
+    val lengthFieldPrepender = new LengthFieldPrepender(4)
 
-  // There are two error handlers by design. WriteErrorHandler adds a future listener to make sure writes to network
-  // succeed. It is added to the head of pipeline (it's the closest of the two to actual network), because some writes
-  // are initiated from the middle of the pipeline (e.g. extension requests). FatalErrorHandler, on the other hand,
-  // reacts to inbound exceptions (the ones thrown during channelRead). It is added to the tail of pipeline to handle
-  // exceptions bubbling up from all the handlers below. When a fatal exception is caught (like OutOfMemory), the
-  // application is terminated.
-  private val writeErrorHandler = new WriteErrorHandler
-  private val fatalErrorHandler = new FatalErrorHandler
-  private val historyReplier = new HistoryReplier(history, settings.synchronizationSettings)
-  private val inboundConnectionFilter: PipelineInitializer.HandlerWrapper = new InboundConnectionFilter(peerDatabase,
-    settings.networkSettings.maxInboundConnections,
-    settings.networkSettings.maxConnectionsPerHost)
+    // There are two error handlers by design. WriteErrorHandler adds a future listener to make sure writes to network
+    // succeed. It is added to the head of pipeline (it's the closest of the two to actual network), because some writes
+    // are initiated from the middle of the pipeline (e.g. extension requests). FatalErrorHandler, on the other hand,
+    // reacts to inbound exceptions (the ones thrown during channelRead). It is added to the tail of pipeline to handle
+    // exceptions bubbling up from all the handlers below. When a fatal exception is caught (like OutOfMemory), the
+    // application is terminated.
+    val writeErrorHandler = new WriteErrorHandler
+    val fatalErrorHandler = new FatalErrorHandler
 
-  private val knownInvalidBlocks = new InvalidBlockStorageImpl(settings.synchronizationSettings.invalidBlocksStorage)
+    val historyReplier = new HistoryReplier(history, settings.synchronizationSettings)
+    val inboundConnectionFilter: PipelineInitializer.HandlerWrapper = new InboundConnectionFilter(peerDatabase,
+      settings.networkSettings.maxInboundConnections,
+      settings.networkSettings.maxConnectionsPerHost)
 
-  private val closedChannels = ConcurrentSubject.publish[Channel](Scheduler.singleThread("closed-channels"))
+    val closedChannelsSubject = ConcurrentSubject.publish[Channel](Scheduler.singleThread("closed-channels"))
+    val (mesageObserver, newtworkMessages) = MessageObserver()
+    val discardingHandler = new DiscardingHandler(lastBlockInfos.map(_.ready))
+    val peerConnections = new ConcurrentHashMap[PeerKey, Channel](10, 0.9f, 10)
+    val serverHandshakeHandler = new HandshakeHandler.Server(handshake, peerInfo, peerConnections, peerDatabase, allChannels)
+    val utxPoolSynchronizer = new UtxPoolSynchronizer(utxPool, allChannels)
 
-  private val (mesageObserver, signatures, blocks, remoteScores, checkpoints, microInvs, micros) = MessageObserver()
-  private val syncWith = RxScoreObserver(settings.synchronizationSettings.scoreTTL, history.score(), blockchainUpdater.lastBlockInfo.map(_.score), remoteScores, closedChannels)
-  private val microblockDatas = MicroBlockSynchronizer(settings.synchronizationSettings.microBlockSynchronizer, history, peerDatabase)(blockchainUpdater.lastBlockInfo.map(_.id), microInvs, micros)
-  private val (extensions, newBlocks) = RxExtensionLoader(settings.synchronizationSettings, history, peerDatabase, knownInvalidBlocks, syncWith, blocks, signatures, closedChannels)
-  private val sink: Observable[Unit] = CoordinatorHandler(checkpointService, history, blockchainUpdater, time,
-    stateReader, utxPool, miner, settings, peerDatabase, allChannels, featureProvider, knownInvalidBlocks)(newBlocks, checkpoints, extensions, microblockDatas)
-  private val discardingHandler = new DiscardingHandler(blockchainUpdater.lastBlockInfo.map(_.ready))
+    val peerSynchronizer: ChannelHandlerAdapter = {
+      if (settings.networkSettings.enablePeersExchange) {
+        new PeerSynchronizer(peerDatabase, settings.networkSettings.peersBroadcastInterval)
+      } else PeerSynchronizer.Disabled
+    }
 
-  private val peerConnections = new ConcurrentHashMap[PeerKey, Channel](10, 0.9f, 10)
+    val serverChannel = settings.networkSettings.declaredAddress.map { _ =>
+      new ServerBootstrap()
+        .group(bossGroup, workerGroup)
+        .channel(classOf[NioServerSocketChannel])
+        .childHandler(new PipelineInitializer[SocketChannel](Seq(
+          inboundConnectionFilter,
+          new HandshakeDecoder(peerDatabase),
+          new HandshakeTimeoutHandler(settings.networkSettings.handshakeTimeout),
+          serverHandshakeHandler,
+          lengthFieldPrepender,
+          new LengthFieldBasedFrameDecoder(100 * 1024 * 1024, 0, 4, 0, 4),
+          new LegacyFrameCodec(peerDatabase),
+          trafficWatcher,
+          discardingHandler,
+          messageCodec,
+          writeErrorHandler,
+          peerSynchronizer,
+          historyReplier,
+          utxPoolSynchronizer,
+          mesageObserver,
+          fatalErrorHandler)))
+        .bind(settings.networkSettings.bindAddress)
+        .channel()
+    }
 
-  private val serverHandshakeHandler = new HandshakeHandler.Server(handshake, peerInfo, peerConnections, peerDatabase, allChannels)
+    val outgoingChannels = new ConcurrentHashMap[InetSocketAddress, Channel]
 
-  private val utxPoolSynchronizer = new UtxPoolSynchronizer(utxPool, allChannels)
+    val clientHandshakeHandler = new HandshakeHandler.Client(handshake, peerInfo, peerConnections, peerDatabase, allChannels)
 
-  sink.subscribe()(monix.execution.Scheduler.Implicits.global)
-
-
-  private val serverChannel = settings.networkSettings.declaredAddress.map { _ =>
-    new ServerBootstrap()
-      .group(bossGroup, workerGroup)
-      .channel(classOf[NioServerSocketChannel])
-      .childHandler(new PipelineInitializer[SocketChannel](Seq(
-        inboundConnectionFilter,
+    val bootstrap = new Bootstrap()
+      .option(ChannelOption.CONNECT_TIMEOUT_MILLIS, settings.networkSettings.connectionTimeout.toMillis.toInt: Integer)
+      .group(workerGroup)
+      .channel(classOf[NioSocketChannel])
+      .handler(new PipelineInitializer[SocketChannel](Seq(
         new HandshakeDecoder(peerDatabase),
         new HandshakeTimeoutHandler(settings.networkSettings.handshakeTimeout),
-        serverHandshakeHandler,
+        clientHandshakeHandler,
         lengthFieldPrepender,
         new LengthFieldBasedFrameDecoder(100 * 1024 * 1024, 0, 4, 0, 4),
         new LegacyFrameCodec(peerDatabase),
@@ -122,125 +145,100 @@ class NetworkServer(checkpointService: CheckpointService,
         utxPoolSynchronizer,
         mesageObserver,
         fatalErrorHandler)))
-      .bind(settings.networkSettings.bindAddress)
-      .channel()
-  }
 
-  private val outgoingChannels = new ConcurrentHashMap[InetSocketAddress, Channel]
+    def formatOutgoingChannelEvent(channel: Channel, event: String) = s"${id(channel)} $event, outgoing channel count: ${outgoingChannels.size()}"
 
-  private val clientHandshakeHandler =
-    new HandshakeHandler.Client(handshake, peerInfo, peerConnections, peerDatabase, allChannels)
+    def handleChannelClosed(remoteAddress: InetSocketAddress)(closeFuture: ChannelFuture): Unit = {
+      closedChannelsSubject.onNext(closeFuture.channel())
+      outgoingChannels.remove(remoteAddress, closeFuture.channel())
+      if (!shutdownInitiated) peerDatabase.suspend(remoteAddress.getAddress)
 
-  private val bootstrap = new Bootstrap()
-    .option(ChannelOption.CONNECT_TIMEOUT_MILLIS, settings.networkSettings.connectionTimeout.toMillis.toInt: Integer)
-    .group(workerGroup)
-    .channel(classOf[NioSocketChannel])
-    .handler(new PipelineInitializer[SocketChannel](Seq(
-      new HandshakeDecoder(peerDatabase),
-      new HandshakeTimeoutHandler(settings.networkSettings.handshakeTimeout),
-      clientHandshakeHandler,
-      lengthFieldPrepender,
-      new LengthFieldBasedFrameDecoder(100 * 1024 * 1024, 0, 4, 0, 4),
-      new LegacyFrameCodec(peerDatabase),
-      trafficWatcher,
-      discardingHandler,
-      messageCodec,
-      writeErrorHandler,
-      peerSynchronizer,
-      historyReplier,
-      utxPoolSynchronizer,
-      mesageObserver,
-      fatalErrorHandler)))
+      if (closeFuture.isSuccess)
+        log.trace(formatOutgoingChannelEvent(closeFuture.channel(), "Channel closed (expected)"))
+      else
+        log.debug(formatOutgoingChannelEvent(closeFuture.channel(), s"Channel closed: ${Option(closeFuture.cause()).map(_.getMessage).getOrElse("no message")}"))
 
-  private val connectTask = workerGroup.scheduleWithFixedDelay(1.second, 5.seconds) {
-    import scala.collection.JavaConverters._
-
-    val outgoing = outgoingChannels.keySet.iterator().asScala.toVector
-
-    def outgoingStr = outgoing.map(_.toString).sorted.mkString("[", ", ", "]")
-
-    val all = peerInfo.values().iterator().asScala.flatMap(_.declaredAddress).toVector
-    val incoming = all.filterNot(outgoing.contains)
-
-    def incomingStr = incoming.map(_.toString).sorted.mkString("[", ", ", "]")
-
-    log.trace(s"Outgoing: $outgoingStr ++ incoming: $incomingStr")
-    if (outgoingChannels.size() < settings.networkSettings.maxOutboundConnections) {
-      peerDatabase
-        .randomPeer(excluded = excludedAddresses ++ all)
-        .foreach(connect)
     }
 
-    Metrics.write(
-      Point
-        .measurement("connections")
-        .addField("outgoing", outgoingStr)
-        .addField("incoming", incomingStr)
-        .addField("n", all.size)
-    )
-  }
-
-  private def peerSynchronizer: ChannelHandlerAdapter = {
-    if (settings.networkSettings.enablePeersExchange) {
-      new PeerSynchronizer(peerDatabase, settings.networkSettings.peersBroadcastInterval)
-    } else PeerSynchronizer.Disabled
-  }
-
-  private def handleChannelClosed(remoteAddress: InetSocketAddress)(closeFuture: ChannelFuture): Unit = {
-    closedChannels.onNext(closeFuture.channel())
-    outgoingChannels.remove(remoteAddress, closeFuture.channel())
-    if (!shutdownInitiated) peerDatabase.suspend(remoteAddress.getAddress)
-
-    if (closeFuture.isSuccess) {
-      log.trace(formatOutgoingChannelEvent(closeFuture.channel(), "Channel closed (expected)"))
-    } else {
-      log.debug(formatOutgoingChannelEvent(
-        closeFuture.channel(),
-        s"Channel closed: ${Option(closeFuture.cause()).map(_.getMessage).getOrElse("no message")}"
-      ))
-    }
-  }
-
-  private def formatOutgoingChannelEvent(channel: Channel, event: String) =
-    s"${id(channel)} $event, outgoing channel count: ${outgoingChannels.size()}"
-
-  private def handleConnectionAttempt(remoteAddress: InetSocketAddress)(thisConnFuture: ChannelFuture): Unit = {
-    if (thisConnFuture.isSuccess) {
-      log.trace(formatOutgoingChannelEvent(thisConnFuture.channel(), "Connection established"))
-      peerDatabase.touch(remoteAddress)
-      thisConnFuture.channel().closeFuture().addListener(handleChannelClosed(remoteAddress))
-    } else if (thisConnFuture.cause() != null) {
-      peerDatabase.suspend(remoteAddress.getAddress)
-      outgoingChannels.remove(remoteAddress, thisConnFuture.channel())
-      thisConnFuture.cause() match {
-        case e: ClosedChannelException =>
-          // this can happen when the node is shut down before connection attempt succeeds
-          log.trace(formatOutgoingChannelEvent(
-            thisConnFuture.channel(),
-            s"Channel closed by connection issue: ${Option(e.getMessage).getOrElse("no message")}"
-          ))
-        case other => log.debug(formatOutgoingChannelEvent(thisConnFuture.channel(), other.getMessage))
+    def handleConnectionAttempt(remoteAddress: InetSocketAddress)(thisConnFuture: ChannelFuture): Unit = {
+      if (thisConnFuture.isSuccess) {
+        log.trace(formatOutgoingChannelEvent(thisConnFuture.channel(), "Connection established"))
+        peerDatabase.touch(remoteAddress)
+        thisConnFuture.channel().closeFuture().addListener(handleChannelClosed(remoteAddress))
+      } else if (thisConnFuture.cause() != null) {
+        peerDatabase.suspend(remoteAddress.getAddress)
+        outgoingChannels.remove(remoteAddress, thisConnFuture.channel())
+        thisConnFuture.cause() match {
+          case e: ClosedChannelException =>
+            // this can happen when the node is shut down before connection attempt succeeds
+            log.trace(formatOutgoingChannelEvent(
+              thisConnFuture.channel(),
+              s"Channel closed by connection issue: ${Option(e.getMessage).getOrElse("no message")}"
+            ))
+          case other => log.debug(formatOutgoingChannelEvent(thisConnFuture.channel(), other.getMessage))
+        }
       }
     }
-  }
 
-  def connect(remoteAddress: InetSocketAddress): Unit =
-    outgoingChannels.computeIfAbsent(remoteAddress, _ => {
-      val newConnFuture = bootstrap.connect(remoteAddress)
 
-      log.trace(s"${id(newConnFuture.channel())} Connecting to $remoteAddress")
-      newConnFuture.addListener(handleConnectionAttempt(remoteAddress)).channel()
-    })
+    def doConnect(remoteAddress: InetSocketAddress): Unit =
+      outgoingChannels.computeIfAbsent(remoteAddress, _ => {
+        val newConnFuture = bootstrap.connect(remoteAddress)
 
-  def shutdown(): Unit = try {
-    shutdownInitiated = true
-    connectTask.cancel(false)
-    serverChannel.foreach(_.close().await())
-    log.debug("Unbound server")
-    allChannels.close().await()
-    log.debug("Closed all channels")
-  } finally {
-    workerGroup.shutdownGracefully().await()
-    bossGroup.shutdownGracefully().await()
+        log.trace(s"${id(newConnFuture.channel())} Connecting to $remoteAddress")
+        newConnFuture.addListener(handleConnectionAttempt(remoteAddress)).channel()
+      })
+
+    val connectTask = workerGroup.scheduleWithFixedDelay(1.second, 5.seconds) {
+      import scala.collection.JavaConverters._
+      val outgoing = outgoingChannels.keySet.iterator().asScala.toVector
+
+      def outgoingStr = outgoing.map(_.toString).sorted.mkString("[", ", ", "]")
+
+      val all = peerInfo.values().iterator().asScala.flatMap(_.declaredAddress).toVector
+      val incoming = all.filterNot(outgoing.contains)
+
+      def incomingStr = incoming.map(_.toString).sorted.mkString("[", ", ", "]")
+
+      log.trace(s"Outgoing: $outgoingStr ++ incoming: $incomingStr")
+      if (outgoingChannels.size() < settings.networkSettings.maxOutboundConnections) {
+        peerDatabase
+          .randomPeer(excluded = excludedAddresses ++ all)
+          .foreach(doConnect)
+      }
+
+      Metrics.write(
+        Point
+          .measurement("connections")
+          .addField("outgoing", outgoingStr)
+          .addField("incoming", incomingStr)
+          .addField("n", all.size)
+      )
+    }
+
+
+
+
+    def doShutdown(): Unit = try {
+      shutdownInitiated = true
+      connectTask.cancel(false)
+      serverChannel.foreach(_.close().await())
+      log.debug("Unbound server")
+      allChannels.close().await()
+      log.debug("Closed all channels")
+    } finally {
+      workerGroup.shutdownGracefully().await()
+      bossGroup.shutdownGracefully().await()
+    }
+
+    new NS {
+      override def connect(remoteAddress: InetSocketAddress): Unit = doConnect(remoteAddress)
+
+      override def shutdown(): Unit = doShutdown()
+
+      override val messages: Messages = newtworkMessages
+
+      override val closedChannels: Observable[Channel] = closedChannelsSubject
+    }
   }
 }
